Refactored recipe processors into simplier design which uses enums rather than hard typing. Added RecipeObjects for items, passive, generator, energy item, and transmutation. Added effiency values for transmtuation recipes. Refactored tile entity machine inventories into modular classes, added support for tile entities which contain conduit interactables. Refactored ItemSlotUI to use prefabs. Added support for ItemSlotUI to display non items

+++ Assets/Scripts/Item/Loot/LootTableHelper.cs
@@ -18,7 +18,7 @@ namespace Items {
                 if (!lootTable.repetitions) {
                     excludedIds.Add(lootResult.item.id);
                 }
-                int itemSlotAmount = Random.Range(lootResult.amountRange.x,lootResult.amountRange.y);
+                uint itemSlotAmount = (uint)Random.Range(lootResult.amountRange.x,lootResult.amountRange.y);
                 itemSlots.Add(new ItemSlot(lootResult.item,itemSlotAmount,null));
             }
             return itemSlots;
